Add Performance evaluation file for September 2018.
These region files did not contain the sceneTagMeta due to which
the predicted labels couldn't be retrieved. Therefore, this file
only contains the Overall Algorithm Accuracy, which is the complement
of the proportion of matchByHand to all the static regions.

The code has been modified to include the predicted labels in the
sceneTagMeta, which will be reflected from subsequent region files.